+++ CHANGELOG.md
@@ -4,6 +4,22 @@ See [UPGRADE.md](UPGRADE.md) for notes on breaking changes for downstream develo
 
 ## Current
 
+## v9.5.0
+
+- Feed close makes the replication detach the channel used for this particular stream.
+
+## v9.4.0
+
+- feed.get accepts onwait hook, that is called if get is waiting for a network peer.
+
+## v9.3.0
+
+- feed.get returns an id, that can be used to cancel a pending get with feed.cancel(getId)
+
+## v9.2.0
+
+- Add `maxBlockSize` on write streams to auto chunk big blocks being written.
+
 ## v9.1.0
 
 - Make peer.remoteOpened public ([#268](https://github.com/hypercore-protocol/hypercore/pull/268))